+++ laktory/cli/__init__.py
@@ -3,5 +3,6 @@ import laktory.cli._init
 import laktory.cli._deploy
 import laktory.cli._preview
 import laktory.cli._quickstart
+import laktory.cli._run
 import laktory.cli._write
 import laktory.cli.app