+++ laktory/cli/_deploy.py
@@ -82,4 +82,3 @@ def deploy(
         controller.terraform_call("apply")
     else:
         raise ValueError(f"backend should be {SUPPORTED_BACKENDS}")
-