+++ laktory/datetime.py
@@ -10,9 +10,9 @@ from planck import units
 
 
 def unix_timestamp(
-        dt: Union[str, int, float, datetime, date] = None,
-        unit: str = "s",
-        as_int: bool = False
+    dt: Union[str, int, float, datetime, date] = None,
+    unit: str = "s",
+    as_int: bool = False,
 ) -> float:
     if dt is None:
         dt = datetime.utcnow()
@@ -46,9 +46,8 @@ def unix_timestamp(
 
 
 def utc_datetime(
-        unixtime: Union[datetime, date, str, float, int] = None,
+    unixtime: Union[datetime, date, str, float, int] = None,
 ) -> datetime:
-
     if not unixtime:
         dt = datetime.utcnow()
     elif isinstance(unixtime, datetime):