+++ tests/test_runner.py
@@ -31,12 +31,15 @@ def test_workspace_client():
 def test_resources():
     dispatcher = Dispatcher(stack=stack)
 
-    assert list(dispatcher.resources.keys()) == ['pl-custom-name', 'job-stock-prices-ut-stack']
+    assert list(dispatcher.resources.keys()) == [
+        "pl-custom-name",
+        "job-stock-prices-ut-stack",
+    ]
     job = dispatcher.resources["job-stock-prices-ut-stack"]
     pl = dispatcher.resources["pl-custom-name"]
 
-    assert job.model_dump() == {'name': 'job-stock-prices-ut-stack', 'id': None}
-    assert pl.model_dump() == {'name': 'pl-stock-prices-ut-stack', 'id': None}
+    assert job.model_dump() == {"name": "job-stock-prices-ut-stack", "id": None}
+    assert pl.model_dump() == {"name": "pl-stock-prices-ut-stack", "id": None}
 
 
 if __name__ == "__main__":